refactor(repl): break up replComponent into focused files
Extract three concerns from the 1500-line repl_pitui.go:

- repl_history.go: replHistory type with load/save/navigate/add.
  Pure data management with file persistence, no UI coupling.
- repl_commands.go: handleCommand dispatcher and sub-commands
  (:help, :env, :type, :find, :doc, etc.)
- repl_completion.go: completion menu show/hide/sync, detail
  bubble positioning, input filtering, and case-sensitive sorting.

The main file (repl_pitui.go) drops from ~1500 to ~760 lines and
retains: struct definition, constructor, input handling (onSubmit,
onKey), eval (startEval, finishEval), overlay components, and the
entry point.

Lock ordering is documented on the struct: r.mu before ev.mu.
