+++ cmd/dang/repl_pitui.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"strings"
 	"sync"
 	"syscall"
@@ -233,6 +232,12 @@ func (d *detailBubble) Render(ctx pitui.RenderContext) pitui.RenderResult {
 
 // ── REPL component ─────────────────────────────────────────────────────────
 
+// replComponent is the main REPL controller. It orchestrates eval, input
+// handling, and UI state. Completion logic lives in repl_completion.go,
+// commands in repl_commands.go, and history in repl_history.go.
+//
+// Lock ordering: r.mu must be acquired before ev.mu. Never hold ev.mu
+// while acquiring r.mu.
 type replComponent struct {
 	mu sync.Mutex
 
@@ -271,9 +276,7 @@ type replComponent struct {
 	daggerLog  *pituiSyncWriter // Dagger log output target (set per-eval)
 
 	// History
-	history      []string
-	historyIndex int
-	historyFile  string
+	history *replHistory
 
 	// Doc browser
 	docBrowser *docBrowserOverlay
@@ -300,8 +303,7 @@ func newReplComponent(ctx context.Context, tui *pitui.TUI, importConfigs []dang.
 		textInput:      ti,
 		entryContainer: &pitui.Container{},
 		menuMaxVisible: 8,
-		historyIndex:   -1,
-		historyFile:    historyFilePath(),
+		history:        newReplHistory(),
 		quit:           make(chan struct{}),
 	}
 
@@ -341,7 +343,7 @@ func newReplComponent(ctx context.Context, tui *pitui.TUI, importConfigs []dang.
 	r.entryContainer.AddChild(ev)
 
 	r.completions = r.buildCompletions()
-	r.loadHistory()
+	r.history.Load()
 
 	return r
 }
@@ -380,7 +382,7 @@ func (r *replComponent) onSubmit(line string) bool {
 		return false
 	}
 
-	r.addHistory(line)
+	r.history.Add(line)
 
 	// Format the echoed input: first line with prompt, continuation lines with "  ... ".
 	inputLines := strings.Split(line, "\n")
@@ -482,12 +484,12 @@ func (r *replComponent) onKey(key uv.Key) bool {
 
 	case key.Code == uv.KeyUp && key.Mod == 0:
 		if !r.menuVisible {
-			r.navigateHistory(-1)
+			r.history.Navigate(-1, r.textInput)
 			return true
 		}
 	case key.Code == uv.KeyDown && key.Mod == 0:
 		if !r.menuVisible {
-			r.navigateHistory(1)
+			r.history.Navigate(1, r.textInput)
 			return true
 		}
 
@@ -634,752 +636,6 @@ func (r *replComponent) finishEval(ev *entryView, logs, results []string, cancel
 	r.inputSlot.Set(r.textInput)
 	r.tui.SetFocus(r.textInput)
 	r.tui.RequestRender(false)
-}
-
-// ── completion menu ─────────────────────────────────────────────────────────
-
-func (r *replComponent) hideCompletionMenu() {
-	r.menuVisible = false
-	r.menuItems = nil
-	r.menuLabels = nil
-	r.menuCompletions = nil
-	if r.menuHandle != nil {
-		r.menuHandle.Hide()
-		r.menuHandle = nil
-		r.menuOverlay = nil
-	}
-	r.hideDetailBubble()
-}
-
-func (r *replComponent) menuDisplayItems() []string {
-	if len(r.menuLabels) > 0 {
-		return r.menuLabels
-	}
-	return r.menuItems
-}
-
-func (r *replComponent) menuBoxWidth() int {
-	items := r.menuDisplayItems()
-	maxW := 0
-	for _, item := range items {
-		if w := lipgloss.Width(item); w > maxW {
-			maxW = w
-		}
-	}
-	if maxW > 60 {
-		maxW = 60
-	}
-	return maxW + 4 // 2 for padding (" item ") + 2 for border
-}
-
-func (r *replComponent) menuBoxHeight() int {
-	n := len(r.menuDisplayItems())
-	h := min(n, r.menuMaxVisible) + 2 // visible items + top/bottom border
-	if n > r.menuMaxVisible {
-		h++ // info line ("1/42")
-	}
-	return h
-}
-
-func (r *replComponent) showCompletionMenu() {
-	xOff := r.completionXOffset()
-	displayItems := r.menuDisplayItems()
-	menuH := r.menuBoxHeight()
-	linesAbove := r.entryContainer.LineCount() // content lines above the input
-
-	var opts *pitui.OverlayOptions
-	if linesAbove >= menuH {
-		// Enough room above: show the menu above the input line.
-		opts = &pitui.OverlayOptions{
-			Width:           pitui.SizeAbs(r.menuBoxWidth()),
-			MaxHeight:       pitui.SizeAbs(menuH),
-			Anchor:          pitui.AnchorBottomLeft,
-			ContentRelative: true,
-			OffsetX:         xOff,
-			OffsetY:         -1, // above the input line
-		}
-	} else {
-		// Not enough room above: show below the input line.
-		// With little content there's no scrolling, so viewport rows
-		// match content rows and we can position without ContentRelative.
-		opts = &pitui.OverlayOptions{
-			Width:     pitui.SizeAbs(r.menuBoxWidth()),
-			MaxHeight: pitui.SizeAbs(menuH),
-			Anchor:    pitui.AnchorTopLeft,
-			Row:       pitui.SizeAbs(linesAbove + 1), // right below the input
-			OffsetX:   xOff,
-		}
-	}
-	if r.menuHandle != nil {
-		// Reuse existing overlay — just update position and data.
-		r.menuOverlay.items = displayItems
-		r.menuOverlay.index = r.menuIndex
-		r.menuOverlay.Update()
-		r.menuHandle.SetOptions(opts)
-	} else {
-		r.menuOverlay = &completionOverlay{
-			items:      displayItems,
-			index:      r.menuIndex,
-			maxVisible: r.menuMaxVisible,
-		}
-		r.menuHandle = r.tui.ShowOverlay(r.menuOverlay, opts)
-	}
-	r.syncDetailBubble()
-}
-
-func (r *replComponent) syncMenu() {
-	if r.menuOverlay != nil {
-		r.menuOverlay.items = r.menuDisplayItems()
-		r.menuOverlay.index = r.menuIndex
-		r.menuOverlay.Update()
-	}
-	r.syncDetailBubble()
-	r.tui.RequestRender(false)
-}
-
-func (r *replComponent) detailBubbleOptions() *pitui.OverlayOptions {
-	xOff := r.completionXOffset()
-	menuW := r.menuBoxWidth()
-	detailX := xOff + menuW + 1 // 1 col gap between menu and detail
-
-	linesAbove := r.entryContainer.LineCount()
-	menuH := min(len(r.menuDisplayItems()), r.menuMaxVisible) + 2
-
-	// Match the menu's vertical positioning strategy.
-	if linesAbove >= menuH {
-		// Menu is above input — detail anchors to same region.
-		return &pitui.OverlayOptions{
-			Width:           pitui.SizePct(35),
-			MaxHeight:       pitui.SizePct(80),
-			Anchor:          pitui.AnchorBottomLeft,
-			ContentRelative: true,
-			OffsetX:         detailX,
-			OffsetY:         -1,
-		}
-	}
-	// Menu is below input.
-	return &pitui.OverlayOptions{
-		Width:     pitui.SizePct(35),
-		MaxHeight: pitui.SizePct(80),
-		Anchor:    pitui.AnchorTopLeft,
-		Row:       pitui.SizeAbs(linesAbove + 1),
-		OffsetX:   detailX,
-	}
-}
-
-func (r *replComponent) showDetailBubble() {
-	opts := r.detailBubbleOptions()
-	if r.detailBubble == nil {
-		r.detailBubble = &detailBubble{}
-		r.detailHandle = r.tui.ShowOverlay(r.detailBubble, opts)
-	} else {
-		r.detailHandle.SetOptions(opts)
-	}
-}
-
-func (r *replComponent) hideDetailBubble() {
-	if r.detailHandle != nil {
-		r.detailHandle.Hide()
-		r.detailHandle = nil
-		r.detailBubble = nil
-	}
-}
-
-func (r *replComponent) syncDetailBubble() {
-	if !r.menuVisible || len(r.menuCompletions) == 0 {
-		r.hideDetailBubble()
-		return
-	}
-	idx := r.menuIndex
-	if idx < 0 || idx >= len(r.menuCompletions) {
-		r.hideDetailBubble()
-		return
-	}
-	c := r.menuCompletions[idx]
-
-	item, found := docItemFromEnv(r.typeEnv, c.Label)
-	if !found {
-		item, found = r.resolveCompletionDocItem(c)
-	}
-	if !found {
-		if c.Detail == "" && c.Documentation == "" {
-			r.hideDetailBubble()
-			return
-		}
-		item = docItem{
-			name:    c.Label,
-			typeStr: c.Detail,
-			doc:     c.Documentation,
-		}
-	}
-
-	r.showDetailBubble()
-	r.detailBubble.item = item
-	r.detailBubble.Update()
-}
-
-// showDetailForCompletion shows the detail bubble for a single completion
-// item, without requiring the dropdown menu to be visible.
-func (r *replComponent) showDetailForCompletion(c dang.Completion) {
-	item, found := docItemFromEnv(r.typeEnv, c.Label)
-	if !found {
-		item, found = r.resolveCompletionDocItem(c)
-	}
-	if !found {
-		if c.Detail == "" && c.Documentation == "" {
-			r.hideDetailBubble()
-			return
-		}
-		item = docItem{
-			name:    c.Label,
-			typeStr: c.Detail,
-			doc:     c.Documentation,
-		}
-	}
-
-	r.showDetailBubble()
-	r.detailBubble.item = item
-	r.detailBubble.Update()
-}
-
-// resolveCompletionDocItem tries to resolve a member completion's docItem
-// by inferring the receiver type from the current input.
-func (r *replComponent) resolveCompletionDocItem(c dang.Completion) (docItem, bool) {
-	val := r.textInput.Value()
-	dotIdx := -1
-	for i := len(val) - 1; i >= 0; i-- {
-		if val[i] == '.' {
-			dotIdx = i
-			break
-		}
-	}
-	if dotIdx < 0 {
-		return docItem{}, false
-	}
-	receiverText := val[:dotIdx]
-	receiverType := dang.InferReceiverType(r.ctx, r.typeEnv, receiverText)
-	if receiverType == nil {
-		return docItem{}, false
-	}
-	unwrapped := unwrapType(receiverType)
-	env, ok := unwrapped.(dang.Env)
-	if !ok {
-		return docItem{}, false
-	}
-	return docItemFromEnv(env, c.Label)
-}
-
-func (r *replComponent) completionXOffset() int {
-	val := r.textInput.Value()
-	// Walk backwards from the cursor to find the start of the completion
-	// token (identifier, possibly preceded by dot+identifier for methods).
-	i := len(val) - 1
-	for i >= 0 && isIdentByte(val[i]) {
-		i--
-	}
-	if i >= 0 && val[i] == '.' {
-		i--
-		for i >= 0 && isIdentByte(val[i]) {
-			i--
-		}
-	}
-	// How many characters from the token start to the cursor end.
-	tokenLen := len(val) - (i + 1)
-	return r.textInput.CursorScreenCol() - tokenLen
-}
-
-func (r *replComponent) updateCompletionMenu() {
-	val := r.textInput.Value()
-
-	if val == "" || strings.HasPrefix(val, ":") {
-		r.hideCompletionMenu()
-		r.textInput.Suggestion = ""
-		return
-	}
-
-	cursorPos := len(val)
-	completions := dang.CompleteInput(r.ctx, r.typeEnv, val, cursorPos)
-
-	if len(completions) > 0 {
-		isArgCompletion := len(completions) > 0 && completions[0].IsArg
-		prefix, partial := splitForSuggestion(val)
-		var matches []string
-		var labels []string
-		var matchCompletions []dang.Completion
-		partialLower := strings.ToLower(partial)
-		for _, c := range completions {
-			cLower := strings.ToLower(c.Label)
-			if cLower == partialLower {
-				continue
-			}
-			if strings.HasPrefix(cLower, partialLower) {
-				if c.IsArg {
-					matches = append(matches, prefix+c.Label+": ")
-					labels = append(labels, c.Label+": "+c.Detail)
-				} else {
-					matches = append(matches, prefix+c.Label)
-					labels = append(labels, c.Label)
-				}
-				matchCompletions = append(matchCompletions, c)
-			}
-		}
-		if !isArgCompletion {
-			matches, matchCompletions = sortByCaseWithCompletions(matches, matchCompletions, prefix, partial)
-			labels, _ = sortByCaseWithCompletions(labels, nil, "", partial)
-		}
-		r.menuLabels = labels
-		r.setMenu(matches, matchCompletions)
-		if len(matches) > 0 {
-			r.textInput.Suggestion = matches[0]
-		} else {
-			r.textInput.Suggestion = ""
-		}
-		return
-	}
-
-	// Fallback: static completions.
-	word := lastIdent(val)
-	if word == "" {
-		r.hideCompletionMenu()
-		r.textInput.Suggestion = ""
-		return
-	}
-
-	var exactCase, otherCase []string
-	wordLower := strings.ToLower(word)
-	for _, c := range r.completions {
-		cLower := strings.ToLower(c)
-		if cLower == wordLower {
-			continue
-		}
-		if strings.HasPrefix(c, word) {
-			exactCase = append(exactCase, c)
-		} else if strings.HasPrefix(cLower, wordLower) {
-			otherCase = append(otherCase, c)
-		}
-	}
-	matches := append(exactCase, otherCase...)
-	r.menuLabels = nil
-	r.setMenu(matches, nil)
-	if len(matches) > 0 {
-		r.textInput.Suggestion = matches[0]
-	} else {
-		r.textInput.Suggestion = ""
-	}
-}
-
-func (r *replComponent) setMenu(matches []string, completions []dang.Completion) {
-	if len(matches) == 0 {
-		r.hideCompletionMenu()
-		return
-	}
-	if len(matches) == 1 {
-		// Single match: no dropdown, but show the detail bubble.
-		r.hideCompletionMenu()
-		if len(completions) == 1 {
-			r.showDetailForCompletion(completions[0])
-		}
-		return
-	}
-	r.menuItems = matches
-	// menuLabels is set by the caller before calling setMenu
-	r.menuCompletions = completions
-	r.menuVisible = true
-	if r.menuIndex >= len(matches) {
-		r.menuIndex = 0
-	}
-	r.showCompletionMenu()
-}
-
-// sortByCaseWithCompletions sorts matches (and their parallel completions)
-// so that exact-case-prefix matches come before case-insensitive ones.
-func sortByCaseWithCompletions(matches []string, completions []dang.Completion, prefix, partial string) ([]string, []dang.Completion) {
-	var exactM, otherM []string
-	var exactC, otherC []dang.Completion
-	for i, m := range matches {
-		suffix := strings.TrimPrefix(m, prefix)
-		if strings.HasPrefix(suffix, partial) {
-			exactM = append(exactM, m)
-			if i < len(completions) {
-				exactC = append(exactC, completions[i])
-			}
-		} else {
-			otherM = append(otherM, m)
-			if i < len(completions) {
-				otherC = append(otherC, completions[i])
-			}
-		}
-	}
-	return append(exactM, otherM...), append(exactC, otherC...)
-}
-
-// ── commands ────────────────────────────────────────────────────────────────
-
-func (r *replComponent) handleCommand(cmdLine string) {
-	r.mu.Lock()
-	ev := r.activeEntryView()
-	ev.mu.Lock()
-	e := ev.entry
-
-	parts := strings.Fields(cmdLine)
-	if len(parts) == 0 {
-		e.writeLogLine(errorStyle.Render("empty command"))
-		ev.Update()
-		ev.mu.Unlock()
-		r.mu.Unlock()
-		return
-	}
-
-	cmd := parts[0]
-	args := parts[1:]
-
-	switch cmd {
-	case "help":
-		e.writeLogLine("Available commands:")
-		e.writeLogLine(dimStyle.Render("  :help      - Show this help"))
-		e.writeLogLine(dimStyle.Render("  :exit      - Exit the REPL"))
-		e.writeLogLine(dimStyle.Render("  :doc       - Interactive API browser"))
-		e.writeLogLine(dimStyle.Render("  :env       - Show environment bindings"))
-		e.writeLogLine(dimStyle.Render("  :type      - Show type of an expression"))
-		e.writeLogLine(dimStyle.Render("  :find      - Find functions/types by pattern"))
-		e.writeLogLine(dimStyle.Render("  :reset     - Reset the environment"))
-		e.writeLogLine(dimStyle.Render("  :clear     - Clear the screen"))
-		e.writeLogLine(dimStyle.Render("  :debug     - Toggle debug mode"))
-		e.writeLogLine(dimStyle.Render("  :debug-render - Toggle render performance logging"))
-		e.writeLogLine(dimStyle.Render("  :version   - Show version info"))
-		e.writeLogLine(dimStyle.Render("  :quit      - Exit the REPL"))
-		e.writeLogLine("")
-		e.writeLogLine(dimStyle.Render("Type Dang expressions to evaluate them."))
-		e.writeLogLine(dimStyle.Render("Tab for completion, Up/Down for history, Alt+Enter for multiline, Ctrl+L to clear."))
-
-	case "exit", "quit":
-		ev.mu.Unlock()
-		r.mu.Unlock()
-		close(r.quit)
-		return
-
-	case "clear":
-		ev.mu.Unlock()
-		r.entryContainer.Clear()
-		r.mu.Unlock()
-		return
-
-	case "reset":
-		r.typeEnv, r.evalEnv = buildEnvFromImports(r.importConfigs)
-		r.refreshCompletions()
-		e.writeLogLine(resultStyle.Render("Environment reset."))
-
-	case "debug":
-		r.debug = !r.debug
-		status := "disabled"
-		if r.debug {
-			status = "enabled"
-		}
-		e.writeLogLine(resultStyle.Render(fmt.Sprintf("Debug mode %s.", status)))
-
-	case "debug-render":
-		r.debugRender = !r.debugRender
-		if r.debugRender {
-			logPath := "/tmp/dang_render_debug.log"
-			f, err := os.OpenFile(logPath, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0644)
-			if err != nil {
-				e.writeLogLine(errorStyle.Render(fmt.Sprintf("failed to open debug log: %v", err)))
-				r.debugRender = false
-			} else {
-				r.debugRenderFile = f
-				r.tui.SetDebugWriter(f)
-				e.writeLogLine(resultStyle.Render(fmt.Sprintf("Render debug enabled. Logging to %s", logPath)))
-				e.writeLogLine(dimStyle.Render("  Use 'tail -f " + logPath + "' in another terminal to watch."))
-			}
-		} else {
-			r.tui.SetDebugWriter(nil)
-			if r.debugRenderFile != nil {
-				r.debugRenderFile.Close()
-				r.debugRenderFile = nil
-			}
-			e.writeLogLine(resultStyle.Render("Render debug disabled."))
-		}
-
-	case "env":
-		r.envCommand(e, args)
-
-	case "version":
-		e.writeLogLine(resultStyle.Render("Dang REPL v0.1.0"))
-		if len(r.importConfigs) > 0 {
-			var names []string
-			for _, ic := range r.importConfigs {
-				names = append(names, ic.Name)
-			}
-			e.writeLogLine(dimStyle.Render(fmt.Sprintf("Imports: %s", strings.Join(names, ", "))))
-		} else {
-			e.writeLogLine(dimStyle.Render("No imports configured (create a dang.toml)"))
-		}
-
-	case "type":
-		r.typeCommand(e, args)
-
-	case "find", "search":
-		r.findCommand(e, args)
-
-	case "history":
-		e.writeLogLine("Recent history:")
-		start := 0
-		if len(r.history) > 20 {
-			start = len(r.history) - 20
-		}
-		for i := start; i < len(r.history); i++ {
-			e.writeLogLine(dimStyle.Render(fmt.Sprintf("  %d: %s", i+1, r.history[i])))
-		}
-
-	case "doc":
-		ev.Update()
-		ev.mu.Unlock()
-		r.mu.Unlock()
-		r.showDocBrowser()
-		return
-
-	default:
-		e.writeLogLine(errorStyle.Render(fmt.Sprintf("unknown command: %s (type :help for available commands)", cmd)))
-	}
-	ev.Update()
-	ev.mu.Unlock()
-	r.mu.Unlock()
-}
-
-func (r *replComponent) envCommand(e *replEntry, args []string) {
-	filter := ""
-	showAll := false
-	if len(args) > 0 {
-		if args[0] == "all" {
-			showAll = true
-		} else {
-			filter = args[0]
-		}
-	}
-	e.writeLogLine("Current environment bindings:")
-	e.writeLogLine("")
-	count := 0
-	for name, scheme := range r.typeEnv.Bindings(dang.PublicVisibility) {
-		if filter != "" && !strings.Contains(strings.ToLower(name), strings.ToLower(filter)) {
-			continue
-		}
-		if !showAll && count >= 20 {
-			e.writeLogLine(dimStyle.Render("  ... use ':env all' to see all"))
-			break
-		}
-		t, _ := scheme.Type()
-		if t != nil {
-			e.writeLogLine(dimStyle.Render(fmt.Sprintf("  %s : %s", name, t)))
-		} else {
-			e.writeLogLine(dimStyle.Render(fmt.Sprintf("  %s", name)))
-		}
-		count++
-	}
-	e.writeLogLine("")
-	e.writeLogLine(dimStyle.Render("Use ':doc' for interactive API browsing"))
-}
-
-func (r *replComponent) typeCommand(e *replEntry, args []string) {
-	if len(args) == 0 {
-		e.writeLogLine(dimStyle.Render("Usage: :type <expression>"))
-		return
-	}
-	expr := strings.Join(args, " ")
-	result, err := dang.Parse("type-check", []byte(expr))
-	if err != nil {
-		e.writeLogLine(errorStyle.Render(fmt.Sprintf("parse error: %v", err)))
-		return
-	}
-	node := result.(*dang.Block)
-	inferredType, err := dang.Infer(r.ctx, r.typeEnv, node, false)
-	if err != nil {
-		e.writeLogLine(errorStyle.Render(fmt.Sprintf("type error: %v", err)))
-		return
-	}
-	e.writeLogLine(fmt.Sprintf("Expression: %s", expr))
-	e.writeLogLine(resultStyle.Render(fmt.Sprintf("Type: %s", inferredType)))
-	trimmed := strings.TrimSpace(expr)
-	if !strings.Contains(trimmed, " ") {
-		if scheme, found := r.typeEnv.SchemeOf(trimmed); found {
-			if t, _ := scheme.Type(); t != nil {
-				e.writeLogLine(dimStyle.Render(fmt.Sprintf("Scheme: %s", scheme)))
-			}
-		}
-	}
-}
-
-func (r *replComponent) findCommand(e *replEntry, args []string) {
-	if len(args) == 0 {
-		e.writeLogLine(dimStyle.Render("Usage: :find <pattern>"))
-		return
-	}
-	pattern := strings.ToLower(args[0])
-	e.writeLogLine(fmt.Sprintf("Searching for '%s'...", pattern))
-	found := false
-	for name, scheme := range r.typeEnv.Bindings(dang.PublicVisibility) {
-		if strings.Contains(strings.ToLower(name), pattern) {
-			t, _ := scheme.Type()
-			if t != nil {
-				e.writeLogLine(dimStyle.Render(fmt.Sprintf("  %s : %s", name, t)))
-			} else {
-				e.writeLogLine(dimStyle.Render(fmt.Sprintf("  %s", name)))
-			}
-			found = true
-		}
-	}
-	for name, env := range r.typeEnv.NamedTypes() {
-		if strings.Contains(strings.ToLower(name), pattern) {
-			doc := env.GetModuleDocString()
-			if doc != "" {
-				if len(doc) > 60 {
-					doc = doc[:57] + "..."
-				}
-				e.writeLogLine(dimStyle.Render(fmt.Sprintf("  %s - %s", name, doc)))
-			} else {
-				e.writeLogLine(dimStyle.Render(fmt.Sprintf("  %s", name)))
-			}
-			found = true
-		}
-	}
-	if !found {
-		e.writeLogLine(dimStyle.Render(fmt.Sprintf("No matches found for '%s'", pattern)))
-	}
-}
-
-// ── completions ─────────────────────────────────────────────────────────────
-
-func (r *replComponent) buildCompletions() []string {
-	return buildCompletionList(r.typeEnv)
-}
-
-func (r *replComponent) refreshCompletions() {
-	r.completions = r.buildCompletions()
-}
-
-// ── history ─────────────────────────────────────────────────────────────────
-
-const maxHistoryEntries = 1000
-
-// historyFilePath returns the path to the history file, respecting
-// XDG_DATA_HOME (default ~/.local/share/dang/history).
-func historyFilePath() string {
-	dir := os.Getenv("XDG_DATA_HOME")
-	if dir == "" {
-		home, err := os.UserHomeDir()
-		if err != nil {
-			return "/tmp/dang_history" // last resort fallback
-		}
-		dir = filepath.Join(home, ".local", "share")
-	}
-	return filepath.Join(dir, "dang", "history")
-}
-
-// historyEncode escapes an entry for single-line storage.
-// Newlines become literal \n, backslashes become \\.
-func historyEncode(s string) string {
-	s = strings.ReplaceAll(s, `\`, `\\`)
-	s = strings.ReplaceAll(s, "\n", `\n`)
-	return s
-}
-
-// historyDecode reverses historyEncode.
-func historyDecode(s string) string {
-	var buf strings.Builder
-	buf.Grow(len(s))
-	for i := 0; i < len(s); i++ {
-		if s[i] == '\\' && i+1 < len(s) {
-			switch s[i+1] {
-			case 'n':
-				buf.WriteByte('\n')
-				i++
-			case '\\':
-				buf.WriteByte('\\')
-				i++
-			default:
-				buf.WriteByte(s[i])
-			}
-		} else {
-			buf.WriteByte(s[i])
-		}
-	}
-	return buf.String()
-}
-
-func (r *replComponent) addHistory(line string) {
-	if len(r.history) > 0 && r.history[len(r.history)-1] == line {
-		r.historyIndex = -1
-		return
-	}
-	r.history = append(r.history, line)
-	r.historyIndex = -1
-	r.appendHistory(line)
-}
-
-func (r *replComponent) navigateHistory(direction int) {
-	if len(r.history) == 0 {
-		return
-	}
-	if direction < 0 {
-		if r.historyIndex == -1 {
-			r.historyIndex = len(r.history) - 1
-		} else if r.historyIndex > 0 {
-			r.historyIndex--
-		}
-	} else {
-		if r.historyIndex == -1 {
-			return
-		}
-		r.historyIndex++
-		if r.historyIndex >= len(r.history) {
-			r.historyIndex = -1
-			r.textInput.SetValue("")
-			return
-		}
-	}
-	if r.historyIndex >= 0 && r.historyIndex < len(r.history) {
-		r.textInput.SetValue(r.history[r.historyIndex])
-		r.textInput.CursorEnd()
-	}
-}
-
-func (r *replComponent) loadHistory() {
-	data, err := os.ReadFile(r.historyFile)
-	if err != nil {
-		return
-	}
-	for _, line := range strings.Split(strings.TrimSpace(string(data)), "\n") {
-		line = strings.TrimSpace(line)
-		if line != "" {
-			r.history = append(r.history, historyDecode(line))
-		}
-	}
-	// Truncate on load if file grew too large (amortized).
-	if len(r.history) > maxHistoryEntries {
-		r.history = r.history[len(r.history)-maxHistoryEntries:]
-		r.rewriteHistory()
-	}
-}
-
-// appendHistory appends a single entry to the history file.
-func (r *replComponent) appendHistory(line string) {
-	_ = os.MkdirAll(filepath.Dir(r.historyFile), 0755)
-	f, err := os.OpenFile(r.historyFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		return
-	}
-	defer f.Close()
-	fmt.Fprintln(f, historyEncode(line))
-}
-
-// rewriteHistory rewrites the entire history file (used for truncation).
-func (r *replComponent) rewriteHistory() {
-	_ = os.MkdirAll(filepath.Dir(r.historyFile), 0755)
-	var buf strings.Builder
-	for _, entry := range r.history {
-		buf.WriteString(historyEncode(entry))
-		buf.WriteByte('\n')
-	}
-	_ = os.WriteFile(r.historyFile, []byte(buf.String()), 0644)
 }
 
 // ── doc browser ─────────────────────────────────────────────────────────────
